+++ kinetic/client.py
@@ -36,6 +36,7 @@ class Client(BaseClient):
                 self.update_header(header)
                 # send message synchronously
                 header, value = self.send(header, value)
+            operations._check_status(header)
             return op.parse(header, value)
         except Exception as e:
             return op.onError(e)